The ANR is avoided by changing the cleanupUnusedAccountDirectories() method from MainApp.kt to DrawerActivity.kt
